+++ Sources/VaporDocs/Languages.swift
@@ -1,0 +1,840 @@
+import Kiln
+
+let languages: [Language] = [
+    Language(
+        .english,
+        isDefault: true,
+        // Docs-specific strings (not part of Kiln's built-in chrome) looked up
+        // in the navbar/footer templates via `#t("key")`. English is the
+        // default language, so these double as the fallback for any key a
+        // translation below hasn't provided yet.
+        customStrings: [
+            "tagline": "Vapor provides a safe, performant and easy to use foundation to build HTTP servers, backends and APIs in Swift",
+            "joinDiscord": "Join our Discord",
+            "supporters": "Supporters",
+            "frameworkDocs": "Framework Docs",
+            "apiDocs": "API Docs",
+            "frameworkDocsCaption": "Learn how to use Vapor",
+            "apiDocsCaption": "Reference documentation for Vapor",
+            "selectLanguage": "Select language",
+            "selectVersion": "Select documentation version",
+            "selectTheme": "Select theme",
+            "closeMenu": "Close menu",
+        ],
+        localisation: .init(
+            searchPlaceholder: "Quick search",
+            tableOfContentsTitle: "Table of contents"
+        )
+    ),
+
+    Language(
+        .german,
+        siteName: "Vapor Dokumentation",
+        description: "Vapors Dokumentation (Web-Framework für Swift).",
+        navTranslations: [
+            "Advanced": "Erweitert",
+            "Authentication": "Authentifzierung",
+            "Basics": "Grundlagen",
+            "Commands": "Befehle",
+            "Content": "Modelbindung",
+            "Contributing": "Mitwirken",
+            "Contributing Guide": "Leitfaden für Beiträge",
+            "Deploy": "Bereitstellung",
+            "Environment": "Umgebung",
+            "Errors": "Fehlerbehandlung",
+            "Files": "Dateien",
+            "Folder Structure": "Verzeichnis",
+            "Getting Started": "Einführung",
+            "Hello, world": "Hallo Welt",
+            "Install": "Installation",
+            "Logging": "Protokollierung",
+            "Migrations": "Migrationen",
+            "Model": "Models",
+            "Overview": "Übersicht",
+            "Query": "Abfrage",
+            "Relations": "Beziehungen",
+            "Security": "Sicherheit",
+            "Services": "Dienste",
+            "Sessions": "Sitzungen",
+            "Testing": "Testen",
+            "Transactions": "Transaktionen",
+            "Validation": "Validierung",
+            "Welcome": "Begrüßung",
+        ],
+        customStrings: [
+            "tagline": "Vapor bietet eine sichere, performante und einfach zu nutzende Grundlage, um HTTP-Server, Backends und APIs in Swift zu erstellen",
+            "joinDiscord": "Tritt unserem Discord bei",
+            "supporters": "Unterstützer",
+            "frameworkDocs": "Framework-Doku",
+            "apiDocs": "API-Doku",
+            "frameworkDocsCaption": "Lerne, wie du Vapor verwendest",
+            "apiDocsCaption": "Referenzdokumentation für Vapor",
+            "selectLanguage": "Sprache auswählen",
+            "selectVersion": "Dokumentationsversion auswählen",
+            "selectTheme": "Design auswählen",
+            "closeMenu": "Menü schließen",
+        ],
+        localisation: .init(
+            searchPlaceholder: "Suchen",
+            searchNoResults: "Keine Ergebnisse gefunden",
+            tableOfContentsTitle: "Auf dieser Seite",
+            previousPage: "Zurück",
+            nextPage: "Weiter",
+            editPage: "Diese Seite bearbeiten",
+            fallbackTitle: "Übersetzung nicht verfügbar",
+            fallbackMessage: "Diese Seite wurde noch nicht übersetzt, daher wird die Standardsprache angezeigt.",
+            notFoundTitle: "Seite nicht gefunden",
+            notFoundMessage: "Die gesuchte Seite wurde möglicherweise verschoben, umbenannt oder existiert nicht.",
+            notFoundLink: "Zurück zur Startseite",
+            toggleNavigation: "Navigation umschalten",
+            toggleColourScheme: "Farbschema umschalten",
+            store: "Shop",
+            showcase: "Projekte",
+            help: "Hilfe",
+            resources: "Ressourcen",
+            language: "Sprache",
+            theme: "Design",
+            lightTheme: "Hell",
+            darkTheme: "Dunkel"
+        )
+    ),
+
+    Language(
+        .spanish,
+        siteName: "Documentación de Vapor",
+        description: "Documentación de Vapor (framework web para Swift).",
+        navTranslations: [
+            "APNS": "APNS",
+            "Advanced": "Avanzado",
+            "Async": "Asincronía",
+            "Authentication": "Autenticación",
+            "Basics": "Fundamentos",
+            "Client": "Cliente",
+            "Commands": "Comandos",
+            "Content": "Content",
+            "Contributing": "Colaborar",
+            "Contributing Guide": "Guía para Colaborar",
+            "Controllers": "Controladores",
+            "Crypto": "Criptografía",
+            "Custom Tags": "Etiquetas Personalizadas",
+            "Deploy": "Desplegar",
+            "Environment": "Entorno",
+            "Errors": "Errores",
+            "Files": "Ficheros",
+            "Fluent": "Fluent",
+            "Folder Structure": "Estructura de Carpetas",
+            "Getting Started": "Comenzando",
+            "Hello, world": "Hola, mundo",
+            "Install": "Instalación",
+            "JWT": "JWT",
+            "Leaf": "Leaf",
+            "Logging": "Logging",
+            "Middleware": "Middleware",
+            "Migrations": "Migraciones",
+            "Model": "Modelo",
+            "Overview": "Presentación",
+            "Passwords": "Contraseñas",
+            "Query": "Consultas",
+            "Queues": "Colas",
+            "Redis": "Redis",
+            "Relations": "Relaciones",
+            "Release Notes": "Notas de Versiones",
+            "Request": "Solicitud",
+            "Routing": "Routing",
+            "Schema": "Esquema",
+            "Security": "Seguridad",
+            "Server": "Servidor",
+            "Services": "Servicios",
+            "Sessions": "Sesiones",
+            "SwiftPM": "SwiftPM",
+            "Testing": "Testing",
+            "Transactions": "Transacciones",
+            "Upgrading": "Actualizar",
+            "Validation": "Validación",
+            "Version (4.0)": "Versión (4.0)",
+            "WebSockets": "WebSockets",
+            "Welcome": "Bienvenido",
+            "Xcode": "Xcode",
+        ],
+        customStrings: [
+            "tagline": "Vapor ofrece una base segura, eficiente y fácil de usar para crear servidores HTTP, backends y APIs en Swift",
+            "joinDiscord": "Únete a nuestro Discord",
+            "supporters": "Colaboradores",
+            "frameworkDocs": "Documentación del framework",
+            "apiDocs": "Documentación de la API",
+            "frameworkDocsCaption": "Aprende a usar Vapor",
+            "apiDocsCaption": "Documentación de referencia de Vapor",
+            "selectLanguage": "Seleccionar idioma",
+            "selectVersion": "Seleccionar versión de la documentación",
+            "selectTheme": "Seleccionar tema",
+            "closeMenu": "Cerrar el menú",
+        ],
+        localisation: .init(
+            searchPlaceholder: "Buscar",
+            searchNoResults: "No se encontraron resultados",
+            tableOfContentsTitle: "En esta página",
+            previousPage: "Anterior",
+            nextPage: "Siguiente",
+            editPage: "Editar esta página",
+            fallbackTitle: "Traducción no disponible",
+            fallbackMessage: "Esta página aún no se ha traducido, por lo que se muestra el idioma predeterminado.",
+            notFoundTitle: "Página no encontrada",
+            notFoundMessage: "Es posible que la página que buscas se haya movido, cambiado de nombre o que nunca haya existido.",
+            notFoundLink: "Volver a la página de inicio",
+            toggleNavigation: "Alternar la navegación",
+            toggleColourScheme: "Alternar el esquema de color",
+            home: "Inicio",
+            store: "Tienda",
+            showcase: "Proyectos",
+            team: "Equipo",
+            help: "Ayuda",
+            pressKit: "Kit de prensa",
+            community: "Comunidad",
+            resources: "Recursos",
+            language: "Idioma",
+            version: "Versión",
+            theme: "Tema",
+            lightTheme: "Claro",
+            darkTheme: "Oscuro",
+            systemTheme: "Sistema"
+        )
+    ),
+
+    Language(
+        .french,
+        siteName: "Documentation de Vapor",
+        description: "Documentation de Vapor (framework web pour Swift).",
+        navTranslations: [
+            "Advanced": "Avancé",
+            "Async": "Asynchrone",
+            "Authentication": "Authentification",
+            "Basics": "Bases",
+            "Commands": "Commandes",
+            "Content": "Contenu",
+            "Contributing": "Contribuer",
+            "Contributing Guide": "Guide de contribution",
+            "Controllers": "Contrôleurs",
+            "Custom Tags": "Tags customisés",
+            "Errors": "Erreurs",
+            "Deploy": "Deployer",
+            "Environment": "Environement",
+            "Files": "Fichiers",
+            "Folder Structure": "Structure du Dossier",
+            "Getting Started": "Commencer",
+            "Hello, world": "Bonjour, monde",
+            "Install": "Installer",
+            "Migrations": "Migrations",
+            "Overview": "Aperçu",
+            "Passwords": "Mots de passe",
+            "Query": "Requête",
+            "Queues": "Files d'attente",
+            "Relataions": "Relations",
+            "Release Notes": "Notes de Version",
+            "Request": "Requête",
+            "Routing": "Routage",
+            "Schema": "Schema",
+            "Security": "Securité",
+            "Services": "Services",
+            "Sessions": "Sessions",
+            "Testing": "Test",
+            "Transactions": "Transactions",
+            "Upgrading": "Mettre à jour",
+            "Validation": "Validation",
+            "Version (4.0)": "Version (4.0)",
+            "Welcome": "Bienvenue",
+        ],
+        customStrings: [
+            "tagline": "Vapor fournit une base sûre, performante et facile à utiliser pour créer des serveurs HTTP, des backends et des API en Swift",
+            "joinDiscord": "Rejoignez notre Discord",
+            "supporters": "Soutiens",
+            "frameworkDocs": "Documentation du framework",
+            "apiDocs": "Documentation de l'API",
+            "frameworkDocsCaption": "Apprenez à utiliser Vapor",
+            "apiDocsCaption": "Documentation de référence de Vapor",
+            "selectLanguage": "Sélectionner la langue",
+            "selectVersion": "Sélectionner la version de la documentation",
+            "selectTheme": "Sélectionner le thème",
+            "closeMenu": "Fermer le menu",
+        ],
+        localisation: .init(
+            searchPlaceholder: "Rechercher",
+            searchNoResults: "Aucun résultat trouvé",
+            tableOfContentsTitle: "Sur cette page",
+            previousPage: "Précédent",
+            nextPage: "Suivant",
+            editPage: "Modifier cette page",
+            fallbackTitle: "Traduction indisponible",
+            fallbackMessage: "Cette page n'a pas encore été traduite, la langue par défaut est donc affichée.",
+            notFoundTitle: "Page introuvable",
+            notFoundMessage: "La page que vous recherchez a peut-être été déplacée, renommée ou n'a jamais existé.",
+            notFoundLink: "Retour à la page d'accueil",
+            toggleNavigation: "Afficher/masquer la navigation",
+            toggleColourScheme: "Changer de thème de couleur",
+            home: "Accueil",
+            store: "Boutique",
+            showcase: "Vitrine",
+            team: "Équipe",
+            help: "Aide",
+            pressKit: "Kit de presse",
+            community: "Communauté",
+            resources: "Ressources",
+            language: "Langue",
+            theme: "Thème",
+            lightTheme: "Clair",
+            darkTheme: "Sombre",
+            systemTheme: "Système"
+        )
+    ),
+
+    Language(
+        .italian,
+        siteName: "Documentazione di Vapor",
+        description: "Documentazione di Vapor (framework web per Swift).",
+        navTranslations: [
+            "APNS": "APNS",
+            "Advanced": "Avanzate",
+            "Async": "Asincrono",
+            "Authentication": "Autenticazione",
+            "Basics": "Basi",
+            "Client": "Client",
+            "Commands": "Comandi",
+            "Content": "Contenuto",
+            "Contributing": "Contribuire",
+            "Contributing Guide": "Guida alla Contribuzione",
+            "Controllers": "Controller",
+            "Crypto": "Crittografia",
+            "Custom Tags": "Tag Personalizzati",
+            "Deploy": "Deploy",
+            "Environment": "Ambiente",
+            "Errors": "Errori",
+            "Files": "File",
+            "Fluent": "Fluent",
+            "Folder Structure": "Struttura della Cartella",
+            "Getting Started": "Inizio",
+            "Hello, world": "Ciao, mondo",
+            "Install": "Installazione",
+            "JWT": "JWT",
+            "Leaf": "Leaf",
+            "Logging": "Logging",
+            "Middleware": "Middleware",
+            "Migrations": "Migrazioni",
+            "Model": "Modello",
+            "Overview": "Panoramica",
+            "Passwords": "Password",
+            "Query": "Query",
+            "Queues": "Code",
+            "Redis": "Redis",
+            "Relations": "Relazioni",
+            "Release Notes": "Note sulla Versione",
+            "Routing": "Routing",
+            "Schema": "Schema",
+            "Security": "Sicurezza",
+            "Server": "Server",
+            "Services": "Servizi",
+            "Sessions": "Sessioni",
+            "SwiftPM": "SwiftPM",
+            "Testing": "Test",
+            "Transactions": "Transazioni",
+            "Upgrading": "Aggiornamento",
+            "Validation": "Validazione",
+            "Version (4.0)": "Versione (4.0)",
+            "WebSockets": "WebSockets",
+            "Welcome": "Benvenuto",
+            "Xcode": "Xcode",
+        ],
+        customStrings: [
+            "tagline": "Vapor offre una base sicura, performante e facile da usare per creare server HTTP, backend e API in Swift",
+            "joinDiscord": "Unisciti al nostro Discord",
+            "supporters": "Sostenitori",
+            "frameworkDocs": "Documentazione del Framework",
+            "apiDocs": "Documentazione API",
+            "frameworkDocsCaption": "Impara a usare Vapor",
+            "apiDocsCaption": "Documentazione di riferimento per Vapor",
+            "selectLanguage": "Seleziona la lingua",
+            "selectVersion": "Seleziona la versione della documentazione",
+            "selectTheme": "Seleziona il tema",
+            "closeMenu": "Chiudi il menu",
+        ],
+        localisation: .init(
+            searchPlaceholder: "Cerca",
+            searchNoResults: "Nessun risultato trovato",
+            tableOfContentsTitle: "In questa pagina",
+            previousPage: "Precedente",
+            nextPage: "Successivo",
+            editPage: "Modifica questa pagina",
+            fallbackTitle: "Traduzione non disponibile",
+            fallbackMessage: "Questa pagina non è ancora stata tradotta, pertanto viene mostrata la lingua predefinita.",
+            notFoundTitle: "Pagina non trovata",
+            notFoundMessage: "La pagina che stai cercando potrebbe essere stata spostata, rinominata o non essere mai esistita.",
+            notFoundLink: "Torna alla home page",
+            toggleNavigation: "Attiva/disattiva la navigazione",
+            toggleColourScheme: "Cambia combinazione di colori",
+            store: "Shop",
+            showcase: "Vetrina",
+            help: "Aiuto",
+            resources: "Risorse",
+            language: "Lingua",
+            version: "Versione",
+            theme: "Tema",
+            lightTheme: "Chiaro",
+            darkTheme: "Scuro",
+            systemTheme: "Sistema"
+        )
+    ),
+
+    Language(
+        .japanese,
+        siteName: "Vapor ドキュメント",
+        description: "Vaporのドキュメント（Swift用Webフレームワーク）。",
+        navTranslations: [
+            "Advanced": "上級",
+            "Async": "非同期",
+            "Authentication": "認証",
+            "Basics": "基礎",
+            "Client": "クライアント",
+            "Commands": "コマンド",
+            "Content": "コンテンツ",
+            "Contributing": "貢献",
+            "Contributing Guide": "貢献ガイド",
+            "Controllers": "コントローラー",
+            "Crypto": "暗号",
+            "Custom Tags": "カスタムタグ",
+            "Deploy": "デプロイ",
+            "Environment": "環境",
+            "Errors": "エラー",
+            "Files": "ファイル",
+            "Folder Structure": "フォルダ構造",
+            "Getting Started": "はじめに",
+            "Install": "インストール",
+            "Logging": "ロギング",
+            "Migrations": "マイグレーション",
+            "Model": "モデル",
+            "Overview": "概要",
+            "Passwords": "パスワード",
+            "Query": "クエリ",
+            "Queues": "キュー",
+            "Relations": "関係",
+            "Release Notes": "リリースノート",
+            "Routing": "ルーティング",
+            "Schema": "スキーマ",
+            "Security": "セキュリティ",
+            "Services": "サービス",
+            "Sessions": "セッション",
+            "Testing": "テスト",
+            "Transactions": "トランザクション",
+            "Upgrading": "アップグレード",
+            "Validation": "バリデーション",
+            "Welcome": "ようこそ",
+        ],
+        customStrings: [
+            "tagline": "Vaporは、SwiftでHTTPサーバー、バックエンド、APIを構築するための、安全で高性能、かつ使いやすい基盤を提供します",
+            "joinDiscord": "Discordに参加",
+            "supporters": "サポーター",
+            "frameworkDocs": "フレームワークドキュメント",
+            "apiDocs": "APIドキュメント",
+            "frameworkDocsCaption": "Vaporの使い方を学ぶ",
+            "apiDocsCaption": "Vaporのリファレンスドキュメント",
+            "selectLanguage": "言語を選択",
+            "selectVersion": "ドキュメントのバージョンを選択",
+            "selectTheme": "テーマを選択",
+            "closeMenu": "メニューを閉じる",
+        ],
+        localisation: .init(
+            searchPlaceholder: "検索",
+            searchNoResults: "結果が見つかりません",
+            tableOfContentsTitle: "このページの内容",
+            previousPage: "前へ",
+            nextPage: "次へ",
+            editPage: "このページを編集",
+            fallbackTitle: "翻訳がありません",
+            fallbackMessage: "このページはまだ翻訳されていないため、デフォルトの言語で表示されています。",
+            notFoundTitle: "ページが見つかりません",
+            notFoundMessage: "お探しのページは移動または名称変更されたか、存在しない可能性があります。",
+            notFoundLink: "ホームページに戻る",
+            toggleNavigation: "ナビゲーションの切り替え",
+            toggleColourScheme: "配色の切り替え",
+            home: "ホーム",
+            store: "ストア",
+            blog: "ブログ",
+            showcase: "ショーケース",
+            team: "チーム",
+            help: "ヘルプ",
+            pressKit: "プレスキット",
+            community: "コミュニティ",
+            resources: "リソース",
+            language: "言語",
+            version: "バージョン",
+            theme: "テーマ",
+            lightTheme: "ライト",
+            darkTheme: "ダーク",
+            systemTheme: "システム"
+        )
+    ),
+
+    Language(
+        .korean,
+        siteName: "Vapor 문서",
+        description: "Vapor 문서 (Swift용 웹 프레임워크).",
+        navTranslations: [
+            "Advanced": "고급",
+            "Async": "비동기 처리",
+            "Authentication": "인증",
+            "Basics": "기본 사항",
+            "Client": "클라이언트",
+            "Commands": "명령어",
+            "Content": "컨텐츠",
+            "Contributing": "기여하기",
+            "Contributing Guide": "기여 가이드",
+            "Controllers": "컨트롤러",
+            "Crypto": "암호화",
+            "Custom Tags": "사용자 정의 태그",
+            "Deploy": "배포",
+            "Environment": "환경 설정",
+            "Errors": "에러",
+            "Files": "파일",
+            "Folder Structure": "폴더 구조",
+            "Getting Started": "시작하기",
+            "Install": "설치",
+            "Logging": "로깅",
+            "Migrations": "마이그레이션",
+            "Model": "모델",
+            "Overview": "개요",
+            "Passwords": "비밀번호",
+            "Query": "쿼리",
+            "Queues": "대기열",
+            "Relations": "관계",
+            "Routing": "라우팅",
+            "Schema": "스키마",
+            "Security": "보안",
+            "Services": "서비스",
+            "Sessions": "세션",
+            "Testing": "테스트",
+            "Transactions": "트랜잭션",
+            "Upgrading": "업그레이드",
+            "Validation": "유효성 검사",
+            "Version (4.0)": "버전 (4.0)",
+            "WebSockets": "웹소켓",
+            "Welcome": "환영합니다",
+        ],
+        customStrings: [
+            "tagline": "Vapor는 Swift로 HTTP 서버, 백엔드, API를 구축하기 위한 안전하고 빠르며 사용하기 쉬운 기반을 제공합니다",
+            "joinDiscord": "Discord 참여하기",
+            "supporters": "후원자",
+            "frameworkDocs": "프레임워크 문서",
+            "apiDocs": "API 문서",
+            "frameworkDocsCaption": "Vapor 사용법 알아보기",
+            "apiDocsCaption": "Vapor 레퍼런스 문서",
+            "selectLanguage": "언어 선택",
+            "selectVersion": "문서 버전 선택",
+            "selectTheme": "테마 선택",
+            "closeMenu": "메뉴 닫기",
+        ],
+        localisation: .init(
+            searchPlaceholder: "검색",
+            searchNoResults: "결과를 찾을 수 없습니다",
+            tableOfContentsTitle: "이 페이지에서",
+            previousPage: "이전",
+            nextPage: "다음",
+            editPage: "이 페이지 편집",
+            fallbackTitle: "번역 없음",
+            fallbackMessage: "이 페이지는 아직 번역되지 않아 기본 언어로 표시됩니다.",
+            notFoundTitle: "페이지를 찾을 수 없습니다",
+            notFoundMessage: "찾고 있는 페이지가 이동되었거나 이름이 변경되었거나 존재하지 않을 수 있습니다.",
+            notFoundLink: "홈페이지로 돌아가기",
+            toggleNavigation: "내비게이션 전환",
+            toggleColourScheme: "색 구성표 전환",
+            home: "홈",
+            store: "스토어",
+            blog: "블로그",
+            showcase: "쇼케이스",
+            team: "팀",
+            help: "도움말",
+            pressKit: "프레스 킷",
+            community: "커뮤니티",
+            resources: "리소스",
+            language: "언어",
+            version: "버전",
+            theme: "테마",
+            lightTheme: "라이트",
+            darkTheme: "다크",
+            systemTheme: "시스템"
+        )
+    ),
+
+    Language(
+        .dutch,
+        siteName: "Vapor Documentatie",
+        description: "Vapor documentatie (webframework voor Swift).",
+        navTranslations: [
+            "Advanced": "Geavanceerd",
+            "Async": "Asynchroon",
+            "Authentication": "Authenticatie",
+            "Basics": "Basis",
+            "Commands": "Commando's",
+            "Content": "Inhoud",
+            "Contributing": "Bijdragen",
+            "Contributing Guide": "Gids Bijdragen",
+            "Crypto": "Encryptie",
+            "Custom Tags": "Zelfgemaakte Tags",
+            "Deploy": "Opzetten",
+            "Environment": "Omgeving",
+            "Files": "Bestanden",
+            "Folder Structure": "Folder Structuur",
+            "Getting Started": "Aan De Slag",
+            "Hello, world": "Hallo, wereld",
+            "Install": "Installeren",
+            "Logging": "Loggen",
+            "Migrations": "Migraties",
+            "Overview": "Overzicht",
+            "Passwords": "Wachtwoorden",
+            "Query": "Opvragen",
+            "Queues": "Wachtrijen",
+            "Relations": "Relaties",
+            "Routing": "Routering",
+            "Schema": "Schema",
+            "Security": "Veiligheid",
+            "Services": "Diensten",
+            "Sessions": "Sessies",
+            "Testing": "Testen",
+            "Transactions": "Transacties",
+            "Upgrading": "Upgraden",
+            "Validation": "Validatie",
+            "Version (4.0)": "Versie (4.0)",
+            "Welcome": "Welkom",
+        ],
+        customStrings: [
+            "tagline": "Vapor biedt een veilige, performante en gebruiksvriendelijke basis om HTTP-servers, backends en API's in Swift te bouwen",
+            "joinDiscord": "Word lid van onze Discord",
+            "supporters": "Ondersteuners",
+            "frameworkDocs": "Framework-documentatie",
+            "apiDocs": "API-documentatie",
+            "frameworkDocsCaption": "Leer hoe je Vapor gebruikt",
+            "apiDocsCaption": "Referentiedocumentatie voor Vapor",
+            "selectLanguage": "Selecteer taal",
+            "selectVersion": "Selecteer documentatieversie",
+            "selectTheme": "Selecteer thema",
+            "closeMenu": "Menu sluiten",
+        ],
+        localisation: .init(
+            searchPlaceholder: "Zoeken",
+            searchNoResults: "Geen resultaten gevonden",
+            tableOfContentsTitle: "Op deze pagina",
+            previousPage: "Vorige",
+            nextPage: "Volgende",
+            editPage: "Deze pagina bewerken",
+            fallbackTitle: "Vertaling niet beschikbaar",
+            fallbackMessage: "Deze pagina is nog niet vertaald, daarom wordt de standaardtaal weergegeven.",
+            notFoundTitle: "Pagina niet gevonden",
+            notFoundMessage: "De pagina die je zoekt is mogelijk verplaatst, hernoemd of heeft nooit bestaan.",
+            notFoundLink: "Terug naar de startpagina",
+            toggleNavigation: "Navigatie schakelen",
+            toggleColourScheme: "Kleurenschema schakelen",
+            store: "Winkel",
+            showcase: "Etalage",
+            resources: "Bronnen",
+            language: "Taal",
+            version: "Versie",
+            theme: "Thema",
+            lightTheme: "Licht",
+            darkTheme: "Donker",
+            systemTheme: "Systeem"
+        )
+    ),
+
+    Language(
+        .polish,
+        siteName: "Dokumentacja Vapor",
+        description: "Dokumentacja Vapor (framework webowy dla Swift).",
+        navTranslations: [
+            "APNS": "APNS",
+            "Advanced": "Zaawansowane",
+            "Async": "Asynchroniczność",
+            "Authentication": "Autentykacja",
+            "Basics": "Podstawy",
+            "Client": "Klient",
+            "Commands": "Komendy",
+            "Content": "Kontent",
+            "Contributing": "Kontrybucja",
+            "Contributing Guide": "Przewodnik do kontrybucji",
+            "Crypto": "Kryptografia",
+            "Custom Tags": "Własne tagi",
+            "Deploy": "Wdrożenie",
+            "Environment": "Środowisko",
+            "Errors": "Błędy",
+            "Files": "Pliki",
+            "Fluent": "Fluent",
+            "Folder Structure": "Struktura folderów",
+            "Getting Started": "Jak zacząć",
+            "Hello, world": "Witaj, świecie",
+            "Install": "Instalacja",
+            "JWT": "JWT",
+            "Leaf": "Leaf",
+            "Logging": "Logowanie",
+            "Middleware": "Middleware",
+            "Migrations": "Migracje",
+            "Model": "Model",
+            "Overview": "Prezentacja",
+            "Passwords": "Hasła",
+            "Query": "Zapytania",
+            "Queues": "Kolejki",
+            "Redis": "Redis",
+            "Relations": "Relacje",
+            "Release Notes": "Informacja o wersji",
+            "Routing": "Kierowanie ruchem",
+            "Schema": "Schematy",
+            "Security": "Bezpieczeństwo",
+            "Server": "Serwer",
+            "Services": "Serwisy",
+            "Sessions": "Sesje",
+            "SwiftPM": "SwiftPM",
+            "Testing": "Testowanie",
+            "Transactions": "Transakcje",
+            "Upgrading": "Aktualizacja",
+            "Validation": "Walidacja",
+            "Version (4.0)": "Wersja (4.0)",
+            "WebSockets": "WebSockety",
+            "Welcome": "Witaj",
+            "Xcode": "Xcode",
+        ],
+        customStrings: [
+            "tagline": "Vapor zapewnia bezpieczną, wydajną i łatwą w użyciu podstawę do budowania serwerów HTTP, backendów i API w Swift",
+            "joinDiscord": "Dołącz do naszego Discorda",
+            "supporters": "Wspierający",
+            "frameworkDocs": "Dokumentacja frameworka",
+            "apiDocs": "Dokumentacja API",
+            "frameworkDocsCaption": "Naucz się korzystać z Vapor",
+            "apiDocsCaption": "Dokumentacja referencyjna Vapor",
+            "selectLanguage": "Wybierz język",
+            "selectVersion": "Wybierz wersję dokumentacji",
+            "selectTheme": "Wybierz motyw",
+            "closeMenu": "Zamknij menu",
+        ],
+        localisation: .init(
+            searchPlaceholder: "Szukaj",
+            searchNoResults: "Nie znaleziono wyników",
+            tableOfContentsTitle: "Na tej stronie",
+            previousPage: "Poprzednia",
+            nextPage: "Następna",
+            editPage: "Edytuj tę stronę",
+            fallbackTitle: "Tłumaczenie niedostępne",
+            fallbackMessage: "Ta strona nie została jeszcze przetłumaczona, dlatego wyświetlany jest język domyślny.",
+            notFoundTitle: "Nie znaleziono strony",
+            notFoundMessage: "Strona, której szukasz, mogła zostać przeniesiona, zmieniona lub nigdy nie istniała.",
+            notFoundLink: "Powrót do strony głównej",
+            toggleNavigation: "Przełącz nawigację",
+            toggleColourScheme: "Przełącz schemat kolorów",
+            home: "Strona główna",
+            store: "Sklep",
+            showcase: "Realizacje",
+            team: "Zespół",
+            help: "Pomoc",
+            pressKit: "Materiały prasowe",
+            community: "Społeczność",
+            resources: "Zasoby",
+            language: "Język",
+            version: "Wersja",
+            theme: "Motyw",
+            lightTheme: "Jasny",
+            darkTheme: "Ciemny",
+            systemTheme: "Systemowy"
+        )
+    ),
+
+    Language(
+        .chinese,
+        siteName: "Vapor 中文文档",
+        description: "Vapor 文档（Swift Web 框架）。",
+        navTranslations: [
+            "APNS": "苹果推送服务",
+            "Advanced": "进阶",
+            "Async": "异步",
+            "Authentication": "认证",
+            "Basics": "入门",
+            "Client": "客户端",
+            "Commands": "命令",
+            "Content": "内容",
+            "Contributing": "贡献",
+            "Contributing Guide": "贡献指南",
+            "Crypto": "加密",
+            "Custom Tags": "自定义标签",
+            "Deploy": "部署",
+            "Environment": "环境",
+            "Errors": "错误",
+            "Files": "文件",
+            "Fluent": "Fluent",
+            "Folder Structure": "项目结构",
+            "Getting Started": "开始",
+            "Hello, world": "你好世界",
+            "Install": "安装",
+            "JWT": "JWT",
+            "Leaf": "Leaf",
+            "Logging": "日志",
+            "Middleware": "中间件",
+            "Migrations": "迁移",
+            "Model": "模型",
+            "Overview": "概述",
+            "Passwords": "密码",
+            "Query": "查询",
+            "Queues": "队列",
+            "Redis": "Redis",
+            "Relations": "关联",
+            "Routing": "路由",
+            "Schema": "模式",
+            "Security": "安全",
+            "Server": "服务器",
+            "Services": "服务",
+            "Sessions": "会话",
+            "SwiftPM": "Swift 包管理器",
+            "Testing": "测试",
+            "Transactions": "事务",
+            "Validation": "验证",
+            "Version (4.0)": "版本 (4.0)",
+            "WebSockets": "即时通讯",
+            "Welcome": "序言",
+            "Xcode": "Xcode",
+        ],
+        customStrings: [
+            "tagline": "Vapor 提供安全、高性能且易于使用的基础框架，助您用 Swift 构建 HTTP 服务器、后端和 API",
+            "joinDiscord": "加入我们的 Discord",
+            "supporters": "赞助者",
+            "frameworkDocs": "框架文档",
+            "apiDocs": "API 文档",
+            "frameworkDocsCaption": "学习如何使用 Vapor",
+            "apiDocsCaption": "Vapor 参考文档",
+            "selectLanguage": "选择语言",
+            "selectVersion": "选择文档版本",
+            "selectTheme": "选择主题",
+            "closeMenu": "关闭菜单",
+        ],
+        localisation: .init(
+            searchPlaceholder: "搜索",
+            searchNoResults: "未找到结果",
+            tableOfContentsTitle: "本页内容",
+            previousPage: "上一页",
+            nextPage: "下一页",
+            editPage: "编辑此页",
+            fallbackTitle: "暂无翻译",
+            fallbackMessage: "此页面尚未翻译，因此显示默认语言。",
+            notFoundTitle: "未找到页面",
+            notFoundMessage: "您要查找的页面可能已被移动、重命名或从未存在。",
+            notFoundLink: "返回首页",
+            toggleNavigation: "切换导航",
+            toggleColourScheme: "切换配色方案",
+            home: "首页",
+            store: "商店",
+            blog: "博客",
+            showcase: "案例展示",
+            team: "团队",
+            help: "帮助",
+            pressKit: "媒体资源",
+            community: "社区",
+            resources: "资源",
+            language: "语言",
+            version: "版本",
+            theme: "主题",
+            lightTheme: "浅色",
+            darkTheme: "深色",
+            systemTheme: "跟随系统"
+        )
+    ),
+]
